+++ Rudder-Braze/Classes/RudderBrazeFactory.h
@@ -6,7 +6,11 @@
 
 #import <Foundation/Foundation.h>
 #import "RudderBrazeIntegration.h"
-@import Rudder;
+#if defined(__has_include) && __has_include(<Rudder/Rudder.h>)
+#import <Rudder/Rudder.h>
+#else
+#import "Rudder.h"
+#endif
 
 NS_ASSUME_NONNULL_BEGIN